chore(be): tighten dashboard theme mutations
