+++ backend/main/lib/helper/permission_config.ex
@@ -150,6 +150,7 @@ defmodule Helper.PermissionConfig do
       "moderator.set" => %{scope: :context, context: :cms, grant: "moderator.set"},
       "moderator.unset" => %{scope: :context, context: :cms, grant: "moderator.unset"},
       "moderator.update" => %{scope: :context, context: :cms, grant: "moderator.update"},
+      "dashboard.theme.update" => %{scope: :context, context: :cms, grant: "community.update"},
       "dashboard.rss.update" => %{scope: :context, context: :cms, grant: "community.update"},
       "community_tag.create" => %{
         scope: :context,
